Change event handler from ConnectivityChanged to ConnectivityTypeChanged.
The objective of this part of the exercise on https://elearning.xamarin.com/forms/xam150/1-obtain-the-devices-network-capabilities/exercise1/9-display-type (step 4-7) was to detect a change in network type (e.g. change from WiFi to Cellular). In order to detect this change we need to subscribe to the ConnectivityTypeChanged event and not to the ConnectivityChanged event again.

Of course the description on the page (see link above) also needs to be updated.